chore: organizes payload config files so they are a bit easier to read and focus on
