+++ .env.local.example
@@ -1,3 +1,3 @@
 PAYLOAD_SECRET=jawliejfilwajefSEANlawefawfewag349jwgo3gj4w
-MONGODB_URI=mongodb://127.0.0.1:27017/next-payload-3
-# POSTGRES_URI=postgresql://postgres:password123@127.0.0.1:5432/next-payload-3
+POSTGRES_URI=postgresql://postgres:password123@127.0.0.1:5432/next-payload-3
+# MONGODB_URI=mongodb://127.0.0.1:27017/next-payload-3
