+++ payload.config.ts
@@ -2,185 +2,27 @@ import path from 'path'
 import { postgresAdapter } from '@payloadcms/db-postgres'
 import { en } from 'payload/i18n/en'
 import {
-  AlignFeature,
-  BlockquoteFeature,
-  BlocksFeature,
-  BoldFeature,
-  ChecklistFeature,
-  HeadingFeature,
-  IndentFeature,
-  InlineCodeFeature,
-  ItalicFeature,
   lexicalEditor,
   LinkFeature,
-  OrderedListFeature,
   ParagraphFeature,
-  RelationshipFeature,
-  UnorderedListFeature,
   UploadFeature,
 } from '@payloadcms/richtext-lexical'
-import { slateEditor } from '@payloadcms/richtext-slate'
-// import { mongooseAdapter } from '@payloadcms/db-mongodb'
-import { buildConfig } from 'payload/config'
+
+import { buildConfig } from 'payload'
 import sharp from 'sharp'
 import { fileURLToPath } from 'url'
 
 const filename = fileURLToPath(import.meta.url)
 const dirname = path.dirname(filename)
-import type { FieldHook } from 'payload/types'
 
-const format = (val: string): string =>
-  val
-    .replace(/ /g, '-')
-    .replace(/[^\w-/]+/g, '')
-    .toLowerCase()
-
-const formatSlug =
-  (fallback: string): FieldHook =>
-  ({ value, originalDoc, data }) => {
-    if (typeof value === 'string') {
-      return format(value)
-    }
-    const fallbackData = data?.[fallback] || originalDoc?.[fallback]
-
-    if (fallbackData && typeof fallbackData === 'string') {
-      return format(fallbackData)
-    }
-
-    return value
-  }
+import { UsersCollection } from '@/cms/collections/Users'
+import { ProductsCollection } from '@/cms/collections/Products'
+import { PagesCollection } from '@/cms/collections/Pages'
+import { MediaCollection } from '@/cms/collections/Media'
 
 export default buildConfig({
-  // editor: slateEditor({}),
-  editor: lexicalEditor({
-    features: ({ defaultFeatures }) => [
-      ...defaultFeatures,
-      LinkFeature(),
-      UploadFeature(),
-      ParagraphFeature(),
-    ],
-  }),
-  collections: [
-    {
-      slug: 'users',
-      auth: true,
-      access: {
-        delete: () => false,
-        update: () => false,
-      },
-      fields: [],
-    },
-    {
-      slug: 'product',
-      admin: {
-        useAsTitle: 'name',
-      },
-      fields: [
-        {
-          name: 'name',
-          type: 'text',
-          required: true,
-        },
-        {
-          name: 'price',
-          type: 'number',
-          required: true,
-        },
-        {
-          name: 'productImage',
-          type: 'upload',
-          relationTo: 'media', // required
-          required: true,
-        },
-        {
-          name: 'description',
-          type: 'richText',
-          required: true,
-        },
-        {
-          name: 'slug',
-          label: 'Slug',
-          type: 'text',
-          admin: {
-            position: 'sidebar',
-          },
-          hooks: {
-            beforeValidate: [formatSlug('name')],
-          },
-        },
-        {
-          name: 'images',
-          type: 'array',
-          fields: [
-            {
-              name: 'image',
-              type: 'upload',
-              relationTo: 'media',
-            },
-          ],
-        },
-      ],
-    },
-    {
-      slug: 'pages',
-      admin: {
-        useAsTitle: 'title',
-      },
-      fields: [
-        {
-          name: 'title',
-          type: 'text',
-        },
-        {
-          name: 'public', // required
-          type: 'checkbox', // required
-          label: 'Public',
-          defaultValue: false,
-        },
-        {
-          name: 'another', // required
-          type: 'checkbox', // required
-          label: 'Another option',
-          defaultValue: false,
-        },
-        {
-          name: 'content',
-          type: 'richText',
-        },
-      ],
-    },
-    {
-      slug: 'media',
-      upload: true,
-      fields: [
-        {
-          name: 'text',
-          type: 'text',
-        },
-      ],
-    },
-  ],
-  secret: process.env.PAYLOAD_SECRET || '',
-  typescript: {
-    outputFile: path.resolve(dirname, 'payload-types.ts'),
-  },
-  db: postgresAdapter({
-    pool: {
-      connectionString: process.env.POSTGRES_URI || '',
-    },
-  }),
-  // db: mongooseAdapter({
-  //   url: process.env.MONGODB_URI || '',
-  // }),
-
-  /**
-   * Payload can now accept specific translations from 'payload/i18n/en'
-   * This is completely optional and will default to English if not provided
-   */
-  i18n: {
-    supportedLanguages: { en },
-  },
-
+  secret: process.env.PAYLOAD_SECRET || 'set-a-secret-in-your-env',
+  collections: [UsersCollection, ProductsCollection, PagesCollection, MediaCollection],
   admin: {
     autoLogin: {
       email: 'dev@payloadcms.com',
@@ -188,12 +30,28 @@ export default buildConfig({
       prefillOnly: true,
     },
   },
+  // the type of DB you would like to use
+  db: postgresAdapter({
+    pool: {
+      connectionString: process.env.POSTGRES_URI || '',
+    },
+  }),
+  // richText editor
+  editor: lexicalEditor(),
+  typescript: {
+    outputFile: path.resolve(dirname, 'payload-types.ts'),
+  },
+  i18n: {
+    supportedLanguages: { en },
+  },
   async onInit(payload) {
     const existingUsers = await payload.find({
       collection: 'users',
       limit: 1,
     })
 
+    // This is useful for local development
+    // so you do not need to create a first-user every time
     if (existingUsers.docs.length === 0) {
       await payload.create({
         collection: 'users',
@@ -207,8 +65,5 @@ export default buildConfig({
   // Sharp is now an optional dependency -
   // if you want to resize images, crop, set focal point, etc.
   // make sure to install it and pass it to the config.
-
-  // This is temporary - we may make an adapter pattern
-  // for this before reaching 3.0 stable
   sharp,
 })
