+++ src/cms/collections/Media/index.ts
@@ -1,0 +1,12 @@
+import { CollectionConfig } from 'payload'
+
+export const MediaCollection: CollectionConfig = {
+  slug: 'media',
+  upload: true,
+  fields: [
+    {
+      name: 'text',
+      type: 'text',
+    },
+  ],
+}
