+++ src/cms/collections/Products/index.ts
@@ -1,0 +1,74 @@
+import { CollectionConfig, FieldHook } from 'payload'
+
+const format = (val: string): string =>
+  val
+    .replace(/ /g, '-')
+    .replace(/[^\w-/]+/g, '')
+    .toLowerCase()
+
+const formatSlug =
+  (fallback: string): FieldHook =>
+  ({ value, originalDoc, data }) => {
+    if (typeof value === 'string') {
+      return format(value)
+    }
+    const fallbackData = data?.[fallback] || originalDoc?.[fallback]
+
+    if (fallbackData && typeof fallbackData === 'string') {
+      return format(fallbackData)
+    }
+
+    return value
+  }
+
+export const ProductsCollection: CollectionConfig = {
+  slug: 'products',
+  admin: {
+    useAsTitle: 'name',
+  },
+  fields: [
+    {
+      name: 'name',
+      type: 'text',
+      required: true,
+    },
+    {
+      name: 'price',
+      type: 'number',
+      required: true,
+    },
+    {
+      name: 'productImage',
+      type: 'upload',
+      relationTo: 'media', // required property
+      required: true,
+    },
+    {
+      name: 'description',
+      type: 'richText',
+      required: true,
+    },
+    {
+      name: 'slug',
+      label: 'Slug',
+      type: 'text',
+      admin: {
+        position: 'sidebar',
+      },
+      hooks: {
+        beforeValidate: [formatSlug('name')],
+      },
+    },
+    {
+      name: 'images',
+      type: 'array',
+      fields: [
+        {
+          name: 'image',
+          type: 'upload',
+          relationTo: 'media', // required property
+        },
+      ],
+    },
+  ],
+}
