+++ src/cms/collections/Users/index.ts
@@ -1,0 +1,11 @@
+import { CollectionConfig } from 'payload'
+
+export const UsersCollection: CollectionConfig = {
+  slug: 'users',
+  auth: true,
+  access: {
+    delete: () => false,
+    update: () => false,
+  },
+  fields: [],
+}
